crypto-bot/requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
web3==6.15.1
pandas==2.2.0
schedule==1.2.1
aiohttp==3.9.3
requests==2.31.0
